Initial ARM support (#1)
Adds initial support for ARM.

ARM does have hardware counters similar to AMD64, except they are typically not readable from userspace. They can be configured to be, but this isn't super typical - we might add support for that later. At a minimum, it usually requires changes in UBoot but depending on the board it may be even more involved.

This adds the basic support for ARM and appropriate switching macros etc.

Tested Jetson Orin Nano running Jetpack 6.