Initial import of analytics tooling (#2)
This is the initial import of analytics tooling from our internal repo. This is not complete. It has been more or less rewritten to suit a wider audience. More will be gradually added as the plan is to rely only on this version and not have our own copy in-tree.